Update index.js to display projects list, btn icon
To automatically display what projects (todo lists) the user currently
has, index.js has been updated to perform said functionality.

As the projects list is now generated via index.js, the temporary html
that was used to style the look of the projects was removed.

Additionally, the HTML of the 'New Project' button was updated with a
new icon, as the 'Add new icons, illustrations' commit removed the old
icon that the button was using previously.