Add edit functionality for projects, todos
Each project and todo has a 'edit square' icon located at the top of
the project header and within the todo when expanded respectively, but
were not yet functional. This has been rectified so that now clicking
on this icon will bring up a modal that allows the user to edit and
save the values of the selected project or todo.