Add delete functionality for projects, todos
Each project and todo has a 'delete' icon located at the top of
the project header and within the todo when expanded respectively, but
were not yet functional. This has been rectified so that now clicking
on this icon will bring up a modal that gives the option to delete the
selected project or todo.