Add 'save' functionality for notes
Each todo has a notes section in their expanded state. Whilst you were
able to enter anything into the section, there was no mechanism that
saved what was written for the given todo. This has been changed so that
any change made onto the notes section in a todo will be saved.