Add 'checked' todo functionality
Each todo has a checkbox, but did not do anything when a user clicked
on it in in previous builds. This has been changed so that it now shows
the todo item in a 'checked' state, where the todo itself will change in
appearance to stand out from the other todos that have not been checked
off yet.

With this, the temporary HTML that was previously used for the purposes
of styling a todo in its 'checked' state, has been removed.