Add date formatting to todo items
The date value that was shown on each todo item came from the JavaScript
Date object. Whilst it does display the date to the user, readability
was not great. The main-todo-list module was updated so that the
date shown is formatted into one that is much simpler for users to read.

Note that the date values that are created and stored within the todo
objects have not changed; this is purely a change on the user interface.