Refactor renderMain fn for more render control
Previously, the renderMain function was responsible for rendering the
entire main element. However, this approach did not allow for rendering
specific parts of the main element. As such, the renderMain was
refactored so that it kept intact the ability to render out the entire
main element, but rendering the parts of the main element have been
delegated to other functions.