Add PubSub class, refactor & modify all codebase
For the codebase to follow a more 'loosely coupled' design, the
'pub-sub.js' file was added in order to apply the PubSub pattern to the
codebase.

However, with how tightly coupled a lot of the components of the app
were, the entire codebase was massively refactored and modified to
decouple a lot of the code, as well apply the PubSub pattern onto the
codebase.